Pumps abort when errorBuffer is full
The last _process is cancelled if its running and the pump aborts
immediately when the errorBuffer is full.

Consider a group with two pumps, having a small buffer size. The second
pump is connected to the the output of the first one. The first pump uses
writeArrayAsync to write its output buffers in batches. Each batch is double
of the buffer size. This construct will result in a deadlock when the error
buffer fills up and the writing of a batch is just started. The second pump
will stop, so it will never read from the output buffer of the first one, so
the writeArrayAsync will never resolve, hence the _process() will never resolve.
That's why we need to abort the pump instead of pausing in when the errorBuffer
is full.